refactor: update service interfaces and move to Domain.Interfaces
